requirements.txt
google-genai==1.5.0
gradio
pydantic==2.10.6